+++ Changelog.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed crash on Godot 3.2.beta6 when pressing the Edit Palette button.
 - The canvas updates automatically when onion skinning settings change.
 - Fixed a rare crash with straight lines. It was possible that the variable `is_making_line` could be true, even if the line itself has been freed from memory.
+- Fixed issue where undo/redo was not working properly for straight lines that went outside the canvas.
 
 ## [v0.6] - 06-01-2020